+++ services/workflows-service/src/common/utils/sign/sign.ts
@@ -4,6 +4,10 @@ import { createHmac, createHash } from 'node:crypto';
  * Signs a payload with a key.
  */
 export const sign = ({ payload, key }: { payload: unknown; key: string }) => {
+  if (!key) {
+    console.warn('No key provided for signing');
+    return 'UNSIGNED';
+  }
   return createHmac('sha256', key).update(JSON.stringify(payload)).digest('hex');
 };