+++ packages/common/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @ballerine/common
 
+## 0.9.25
+
+### Patch Changes
+
+- version bump
+
+## 0.9.24
+
+### Patch Changes
+
+- merchant screening changes
+
+## 0.9.23
+
+### Patch Changes
+
+- added merchant screening to swagger
+
 ## 0.9.22
 
 ### Patch Changes