+++ packages/react-pdf-toolkit/CHANGELOG.md
@@ -1,5 +1,26 @@
 # @ballerine/react-pdf-toolkit
 
+## 1.2.21
+
+### Patch Changes
+
+- Updated dependencies
+  - @ballerine/ui@0.5.21
+
+## 1.2.20
+
+### Patch Changes
+
+- Updated dependencies
+  - @ballerine/ui@0.5.20
+
+## 1.2.19
+
+### Patch Changes
+
+- Updated dependencies
+  - @ballerine/ui@0.5.19
+
 ## 1.2.18
 
 ### Patch Changes