+++ packages/workflow-core/CHANGELOG.md
@@ -1,5 +1,26 @@
 # @ballerine/workflow-core
 
+## 0.6.36
+
+### Patch Changes
+
+- version bump
+- Updated dependencies
+  - @ballerine/common@0.9.25
+
+## 0.6.35
+
+### Patch Changes
+
+- fixed handling of child workflows
+
+## 0.6.34
+
+### Patch Changes
+
+- Updated dependencies
+  - @ballerine/common@0.9.23
+
 ## 0.6.33
 
 ### Patch Changes